% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopernicusDataspace-package.R
\docType{package}
\name{CopernicusDataspace-package}
\alias{CopernicusDataspace}
\alias{CopernicusDataspace-package}
\title{CopernicusDataspace: Search Download and Handle Data from the Copernicus Data Space Ecosystem}
\description{
The Copernicus Data Space Ecosystem, is an open ecosystem that provides free instant access to a wide range of data and services from the Copernicus Sentinel missions and more on our planet’s land, oceans and atmosphere. This package provides entry points to several APIs allowing users to access the data directly in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pepijn-devries/CopernicusDataspace}
  \item \url{https://pepijn-devries.github.io/CopernicusDataspace/}
  \item Report bugs at \url{https://github.com/pepijn-devries/CopernicusDataspace/issues}
}

}
\author{
\strong{Maintainer}: Pepijn de Vries \email{pepijn.devries@outlook.com} (\href{https://orcid.org/0000-0002-7961-6646}{ORCID})

Other contributors:
\itemize{
  \item LVVN (Ministerie van Landbouw, Visserij, Voedselzekerheid en Natuur) (03b1hdw57) [funder]
  \item WMR (Wageningen Marine Research) (013dzwk66) [funder]
}

}
\keyword{internal}
