% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{dse_get_token_details}
\alias{dse_get_token_details}
\title{Decode Access Token}
\usage{
dse_get_token_details(token = dse_access_token())
}
\arguments{
\item{token}{A token obtained with \code{\link[=dse_access_token]{dse_access_token()}} or
\code{\link[=dse_public_access_token]{dse_public_access_token()}}.}
}
\value{
A named \code{list} with token info
}
\description{
This function decodes an access token and returns a named \code{list} with
information from the token.
}
\examples{
if (interactive() && dse_has_client_info()) {
  dse_get_token_details()
}
}
\seealso{
Other authentication: 
\code{\link{dse_access_token}()},
\code{\link{dse_has_s3_secret}()},
\code{\link{dse_set_gdal_token}()},
\code{\link{dse_usage}()}
}
\concept{authentication}
