% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odata_products.R
\name{dse_odata_attributes}
\alias{dse_odata_attributes}
\title{List OData Attributes}
\usage{
dse_odata_attributes(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
A \code{data.frame} listing all attributes for each collection.
}
\description{
Collect a list of OData attributes that can be used for
filtering products with \code{\link[=dse_odata_products_request]{dse_odata_products_request()}}.
}
\examples{
if (interactive()) {
  dse_odata_attributes()
}
}
\seealso{
Other odata:
\code{\link[=dse_odata_bursts_request]{dse_odata_bursts_request()}},
\code{\link[=dse_odata_download]{dse_odata_download()}},
\code{\link[=dse_odata_download_path]{dse_odata_download_path()}},
\code{\link[=dse_odata_product_nodes]{dse_odata_product_nodes()}},
\code{\link[=dse_odata_products_request]{dse_odata_products_request()}},
\code{\link[=dse_odata_quicklook]{dse_odata_quicklook()}}
}
\concept{odata}
