% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odata_bursts.R
\name{dse_odata_bursts_request}
\alias{dse_odata_bursts_request}
\alias{dse_odata_bursts}
\title{Create a OData Request for a Data Space Ecosystem Bursts Data}
\usage{
dse_odata_bursts_request(...)

dse_odata_bursts(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
Returns a \code{data.frame} with burst information.
}
\description{
Obtain metadata for burst data associated with specific products.
}
\details{
For more details about bursts check the
\href{https://documentation.dataspace.copernicus.eu/APIs/Sentinel-1\%20SLC\%20Burst.html}{burst API documentation}.

You can apply some tidyverse functions (see \link{tidy_verbs}) to \code{odata_request}
object returned by \code{\link[=dse_odata_bursts_request]{dse_odata_bursts_request()}}. These apply lazy evaluation.
Meaning that they are just added to the object and are only evaluated after
calling either \code{\link[dplyr:compute]{dplyr::compute()}} or \code{\link[dplyr:compute]{dplyr::collect()}} (see examples).
}
\examples{
if (interactive()) {
  dse_odata_bursts(ParentProductId == "879d445c-2c67-5b30-8589-b1f478904269")
}
}
