% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odata_products.R
\name{dse_odata_download}
\alias{dse_odata_download}
\title{Download Data Space Ecosystem Products Through OData API}
\usage{
dse_odata_download(
  request,
  destination,
  ...,
  s3_key = dse_s3_key(),
  s3_secret = dse_s3_secret()
)
}
\arguments{
\item{request}{A request containing products or burst data that you wish to download.
Use \code{\link[=dse_odata_products_request]{dse_odata_products_request()}} or \code{\link[=dse_odata_bursts_request]{dse_odata_bursts_request()}} to formulate
product or burst specifications.}

\item{destination}{A \code{character} string specifying the directory path,
where to store downloaded products}

\item{...}{Arguments passed to \code{\link[=dse_s3]{dse_s3()}}.}

\item{s3_key, s3_secret}{The s3 key and secret registered under your Data Space
Ecosystem account}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Use \code{\link[=dse_odata_products]{dse_odata_products()}} or \code{\link[=dse_odata_products_request]{dse_odata_products_request()}},
\code{\link[=dse_odata_bursts]{dse_odata_bursts()}} or \code{\link[=dse_odata_bursts_request]{dse_odata_bursts_request()}} to find
a product or burst information. Use this function to download the product(s)
or burst information.
}
\examples{
if (interactive() && dse_has_s3_secret()) {
  dse_odata_download(
    dse_odata_products(Name == "S1C_AUX_PP2_V20241204T000000_G20251024T110034.SAFE"),
    destination = tempdir())
  dse_odata_download(
    dse_odata_products(
      Name ==
        "S1A_IW_OCN__2SDH_20250707T210608_20250707T210625_059983_07739B_893E.SAFE"),
    destination = tempdir())
  dse_odata_download(
    dse_odata_products(
      Id \%in\% c("c8ed8edb-9bef-4717-abfd-1400a57171a4",
                "86288a07-560c-364f-b8ce-669d95f06fa0")),
    destination = tempdir())
}
}
