% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odata_products.R
\name{dse_odata_product_nodes}
\alias{dse_odata_product_nodes}
\title{List OData Product Nodes (i.e. Files and Directories)}
\usage{
dse_odata_product_nodes(product, node_path = "", recursive = FALSE, ...)
}
\arguments{
\item{product}{A product identifier (\code{Id})}

\item{node_path}{Path of nodes separated by forward slashes (\code{"/"}).
Path for which to list nodes. Default is \code{""}, which is the root of
the product}

\item{recursive}{A \code{logical} value. If set to \code{TRUE}, it will recursively
list all nested nodes. Default is \code{FALSE}.}

\item{...}{Ignored}
}
\value{
A \code{data.frame} with information on the requested node(s)
}
\description{
If you know the product \code{Id}, you can use this function to retrieve
information about nodes (i.e. files and directories) within the product.
}
\examples{
if (interactive()) {
  nodes <- dse_odata_product_nodes("c8ed8edb-9bef-4717-abfd-1400a57171a4")
  nodes <- dse_odata_product_nodes("c8ed8edb-9bef-4717-abfd-1400a57171a4",
                                   recursive = TRUE)
}
}
