% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odata_products.R
\name{dse_odata_quicklook}
\alias{dse_odata_quicklook}
\title{Download a Quicklook for a Product}
\usage{
dse_odata_quicklook(product, destination, ...)
}
\arguments{
\item{product}{Identifier (Id) for the product for which to obtain a quicklook.}

\item{destination}{A destination path where to store the image.}

\item{...}{Ignored}
}
\value{
Returns \code{NULL} invisibly
}
\description{
Downloads a 'quicklook' for a product. If the \code{rstudioapi} package is installed,
it will attempt to open the image in the Viewer panel.
}
\examples{
if (interactive()) {
  dse_odata_quicklook(
    "91822f33-b15c-5b60-aa39-6d9f6f5c773b",
    tempfile(fileext = ".jpg"))
}
}
