% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{dse_s3_uri_to_vsi}
\alias{dse_s3_uri_to_vsi}
\title{Convert Uniform Resource Identifier to Virtual System Identifier}
\usage{
dse_s3_uri_to_vsi(uri, streaming = TRUE)
}
\arguments{
\item{uri}{A Uniform Resource Identifier, pointing to an
S3 storage file. You can retrieve one with \code{\link[=dse_stac_get_uri]{dse_stac_get_uri()}}.}

\item{streaming}{A \code{logical} value that allows to toggle between
\code{"\\\\vsis3\\\\"} and \code{"\\\\vsis3_streaming\\\\"} (default). The latter is
faster for reading files from its resource, but does not allow
random access. The first supports random access, but is not as fast
at reading.}
}
\value{
A \code{character} string representing the VSI
}
\description{
Convert Uniform Resource Identifier (URI) to Virtual System
Identifier (VSI). The Copernicus Data Space Ecosystem returns
URIs for accessing assets. Packages that use the GDAL library
(e.g., \code{stars}, \code{terra} and \code{gdalraster})
can use VSI to access raster data directly. Use this function
to convert such an URI to a VIS.
}
\examples{
if (interactive()) {
  dse_stac_get_uri(
    "S1A_IW_GRDH_1SDV_20241125T055820_20241125T055845_056707_06F55C_12F9_COG",
    "vh", "sentinel-1-grd") |>
    dse_s3_uri_to_vsi()
}
}
\seealso{
Other s3: 
\code{\link{dse_has_s3_secret}()},
\code{\link{dse_s3_download}()},
\code{\link{dse_s3_set_gdal_options}()}
}
\concept{s3}
