% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinelhub.R
\name{dse_sh_collections}
\alias{dse_sh_collections}
\title{List Sentinel Hub Collections}
\usage{
dse_sh_collections(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
Returns a \code{data.frame} with information about the
collections available from the Sentinel Hub
}
\description{
List collections that are available from the Sentinel Hub.
}
\examples{
if (interactive()) {
  dse_sh_collections()
}
}
\seealso{
Other sentinelhub: 
\code{\link{dse_sh_custom_scripts}()},
\code{\link{dse_sh_features}()},
\code{\link{dse_sh_get_custom_script}()},
\code{\link{dse_sh_prepare_input}()},
\code{\link{dse_sh_process}()},
\code{\link{dse_sh_queryables}()},
\code{\link{dse_sh_search_request}()},
\code{\link{dse_sh_use_requests_builder}()}
}
\concept{sentinelhub}
