% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinelhub.R
\name{dse_sh_custom_scripts}
\alias{dse_sh_custom_scripts}
\title{List Custom JavaScripts for Processing Sentinel Hub Data}
\usage{
dse_sh_custom_scripts(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
Returns a \code{data.frame} with custom scripts, containing
a column with a title and one with a relative URL.
}
\description{
Custom Eval Scripts, that can be used in Sentinel Hub requests,
for processing data. This functions lists scripts available
from \url{https://github.com/sentinel-hub/custom-scripts}. They
can be retrieved with \code{\link[=dse_sh_get_custom_script]{dse_sh_get_custom_script()}}.
}
\details{
Make sure that you have sufficient monthly quota left to
process images. You can check with \code{\link[=dse_usage]{dse_usage()}}.
}
\examples{
if (interactive()) {
  dse_sh_custom_scripts()
}
}
\references{
\itemize{
\item \url{https://custom-scripts.sentinel-hub.com/}
\item \url{https://github.com/sentinel-hub/custom-scripts}
}
}
