% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinelhub.R
\name{dse_sh_features}
\alias{dse_sh_features}
\title{List Sentinel Hub Features}
\usage{
dse_sh_features(
  collection,
  bbox,
  datetime,
  limit = 10,
  ...,
  token = dse_access_token()
)
}
\arguments{
\item{collection}{A collection for which to list the features.
See \code{\link[=dse_sh_collections]{dse_sh_collections()}} for a list of Sentinel Hub collections.}

\item{bbox}{An object that can be converted into a \code{bbox} class
object (see \code{\link[sf:st_bbox]{sf::st_bbox()}}).}

\item{datetime}{A date-time object, or a vector of two date time
objects (in case of a range). Or an object that can be converted into
a datetime object.}

\item{limit}{The number of records to which the output is limited.
Should be between 1 and 100, and defaults to 10.}

\item{...}{Ignored}

\item{token}{For authentication, many of the Dataspace Ecosystem uses
an access token. Either provide your access token, or obtain one automatically
with \code{\link[=dse_access_token]{dse_access_token()}} (default). Without a valid token you will likely get
an "access denied" error.}
}
\value{
Returns a \code{data.frame} listing features available on
SentinelHub for processing.
}
\description{
List Sentinel Hub features for a specified period and region.
}
\examples{
if (interactive() && dse_has_client_info()) {
  dse_sh_features(
    collection = "sentinel-2-l2a",
    bbox       = c(5.261, 52.680, 5.319, 52.715),
    datetime   = c("2025-01-01 UTC", "2025-01-07 UTC"))
}
}
