% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinelhub.R
\name{dse_sh_get_custom_script}
\alias{dse_sh_get_custom_script}
\title{Retrieve Custom JavaScripts to be Used by Sentinel Hub}
\usage{
dse_sh_get_custom_script(rel_url)
}
\arguments{
\item{rel_url}{A relative URL found with \code{\link[=dse_sh_custom_scripts]{dse_sh_custom_scripts()}}.}
}
\value{
A \code{character} string containing JavaScript code. This
script can be used with \code{\link[=dse_sh_process]{dse_sh_process()}}
}
\description{
Sentinel Hub uses JavaScripts to process satellite images.
There is a repository with such custom scripts. They can be
listed with \code{\link[=dse_sh_custom_scripts]{dse_sh_custom_scripts()}}. Use the relative URL
(relUrl) from that list to obtain the actual script with this
function.
}
\examples{
if (interactive()) {
  dse_sh_get_custom_script("/sentinel-2/tonemapped_natural_color/")
}
}
