% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinelhub.R
\name{dse_sh_prepare_input}
\alias{dse_sh_prepare_input}
\title{TODO}
\usage{
dse_sh_prepare_input(bounds, data_filter, data_type = "sentinel-2-l2a")
}
\arguments{
\item{bounds}{TODO}

\item{data_filter}{TODO}

\item{data_type}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
library(sf)

shape <- st_bbox(c(xmin = 5.261, ymin = 52.680,
                   xmax = 5.319, ymax = 52.715), crs = 4326) |>
           st_as_sfc()
dse_sh_prepare_input(
  bounds = c(5.261, 52.680, 5.319, 52.715)
)
dse_sh_prepare_input(
  bounds = shape
)
}
\references{
\itemize{
\item \url{https://apps.sentinel-hub.com/requests-builder/}
}
}
