% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinelhub.R
\name{dse_sh_queryables}
\alias{dse_sh_queryables}
\title{List Queryable Fields on Sentinel Hub}
\usage{
dse_sh_queryables(collection, ..., token = dse_access_token())
}
\arguments{
\item{collection}{Collection id for which to obtain queryable fields.}

\item{...}{Ignored.}

\item{token}{For authentication, many of the Dataspace Ecosystem uses
an access token. Either provide your access token, or obtain one automatically
with \code{\link[=dse_access_token]{dse_access_token()}} (default). Without a valid token you will likely get
an "access denied" error.}
}
\value{
Returns a named \code{list}, with information about queryable
fields for the specified \code{collection}.
}
\description{
Return queryable fields for a specific collection on Sentinel Hub.
This is useful information when composing a query with
\code{\link[=dse_sh_prepare_input]{dse_sh_prepare_input()}}. Use \code{\link[=dse_sh_collections]{dse_sh_collections()}} to list
available collections.
}
\examples{
if (interactive() && dse_has_client_info()) {
  dse_sh_queryables("sentinel-2-l1c")
}
}
