% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinelhub.R
\name{dse_sh_use_requests_builder}
\alias{dse_sh_use_requests_builder}
\title{Use Requests Builder to Send Processing Request to SentinelHub}
\usage{
dse_sh_use_requests_builder(
  build,
  destination,
  ...,
  token = dse_access_token()
)
}
\arguments{
\item{build}{A \code{character} string copied from the Request Preview
panel at
\href{https://apps.sentinel-hub.com/requests-builder/}{Sentinel Requests Builder}.
See \code{system.file("requests-builder.txt", package = "CopernicusDataspace")}
for an example of such a text. When you omit this argument, this function
will attempt to retrieve the text from the system's clipboard.}

\item{destination}{A file name to store the downloaded image.}

\item{...}{Ignored}

\item{token}{For authentication, many of the Dataspace Ecosystem uses
an access token. Either provide your access token, or obtain one automatically
with \code{\link[=dse_access_token]{dse_access_token()}} (default). Without a valid token you will likely get
an "access denied" error.}
}
\value{
A \code{httr2_response} class object obtained after sending the request.
}
\description{
Use
\href{https://apps.sentinel-hub.com/requests-builder/}{Sentinel Requests Builder}
to compose a request. Copy the text from the 'Request Preview' panel
and submit with this function. Use \code{\link[=dse_sh_process]{dse_sh_process()}} when you want
to define a request in R, without using a web browser.
}
\examples{
## Read text copied from 'Request Preview' panel on
## <https://apps.sentinel-hub.com/requests-builder/>:
requests_builder <-
  system.file("requests-builder.txt", package = "CopernicusDataspace") |>
    readLines(warn = FALSE) |>
    paste(collapse = "\n")

if (interactive() && dse_has_client_info()) {
  dest <- tempfile(fileext = ".tiff")
  dse_sh_use_requests_builder(requests_builder, destination = dest)
}
}
\references{
\itemize{
\item \url{https://apps.sentinel-hub.com/requests-builder/}
}
}
