% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac.R
\name{dse_stac_api_specs}
\alias{dse_stac_api_specs}
\title{Get STAC API Specifications}
\usage{
dse_stac_api_specs(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
TODO
}
\description{
TODO
}
\examples{
if (interactive()) {
  dse_stac_api_specs()
}
}
