% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac.R
\name{dse_stac_collections}
\alias{dse_stac_collections}
\title{Get a Summary of all Data Space Ecosystem Collections}
\usage{
dse_stac_collections(collection, limit = 1000L, ...)
}
\arguments{
\item{collection}{A specific collection for which to obtain summary information.
If missing (default), all collections are returned.}

\item{limit}{Maximum number of collections to be returned.}

\item{...}{Ignored}
}
\value{
Returns a \code{data.frame} with the requested information
}
\description{
Use the STAC API to get a summary of all collections available
from the interface.
}
\examples{
if (interactive()) {
  dse_stac_collections()
  dse_stac_collections("sentinel-2-l2a")
}
}
\seealso{
Other stac:
\code{\link[=dse_stac_client]{dse_stac_client()}},
\code{\link[=dse_stac_download]{dse_stac_download()}},
\code{\link[=dse_stac_get_uri]{dse_stac_get_uri()}},
\code{\link[=dse_stac_guess_collection]{dse_stac_guess_collection()}},
\code{\link[=dse_stac_queryables]{dse_stac_queryables()}},
\code{\link[=dse_stac_search_request]{dse_stac_search_request()}}
}
\concept{stac}
