% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac.R
\name{dse_stac_download}
\alias{dse_stac_download}
\title{TODO}
\usage{
dse_stac_download(..., destination, token = dse_access_token())
}
\arguments{
\item{...}{TODO}

\item{destination}{TODO}

\item{token}{For authentication, many of the Dataspace Ecosystem uses
an access token. Either provide your access token, or obtain one automatically
with \code{\link[=dse_access_token]{dse_access_token()}} (default). Without a valid token you will likely get
an "access denied" error.}
}
\value{
TODO
}
\description{
TODO
}
\examples{
if (interactive() && dse_has_client_info()) {
  dse_stac_download(destination = tempdir()) #TODO
}
}
