% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac.R
\name{dse_stac_get_uri}
\alias{dse_stac_get_uri}
\title{Get a Uniform Resource Identifier (URI) for an Asset in a Product}
\usage{
dse_stac_get_uri(
  asset_id,
  asset,
  collection = dse_stac_guess_collection,
  type = "s3",
  ...
)
}
\arguments{
\item{asset_id}{STAC id, used for locating the asset download details.}

\item{asset}{Name of the asset to download}

\item{collection}{The identifier for a collection. The default argument
is the \code{\link[=dse_stac_guess_collection]{dse_stac_guess_collection()}} function which tries to guess the
collection id from the \code{asset_id}. A more rigid approach is to provide
the collection id as a \code{character} string.}

\item{type}{Which type of URI should be returned? Defaults
to \code{"s3"}. Use \code{"odata"} to get the alternative https URI.}

\item{...}{Ignored}
}
\value{
Returns the URI as a \code{character} string.
If available, the local path for an asset is returned as attribute.
}
\description{
Get a Uniform Resource Identifier (URI) for an asset in a product.
This can be used to download a file manually or connect to the
asset directly straight from the source.
}
\examples{
if (interactive()) {
  dse_stac_get_uri(
    asset_id = "S2A_MSIL1C_20260109T132741_N0511_R024_T39XVL_20260109T142148",
    asset = "B01"
  )
}
}
