% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac.R
\name{dse_stac_guess_collection}
\alias{dse_stac_guess_collection}
\title{TODO}
\usage{
dse_stac_guess_collection(asset_id)
}
\arguments{
\item{asset_id}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
}
