% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{dse_usage}
\alias{dse_usage}
\alias{dse_user_statistics}
\title{Get Dataspace Account Information}
\usage{
dse_usage(..., token = dse_access_token())

dse_user_statistics(
  range = "DAYS-31",
  resolution = "DAILY",
  token = dse_access_token()
)
}
\arguments{
\item{...}{Ignored}

\item{token}{For authentication, many of the Dataspace Ecosystem uses
an access token. Either provide your access token, or obtain one automatically
with \code{\link[=dse_access_token]{dse_access_token()}} (default). Without a valid token you will likely get
an "access denied" error.}

\item{range}{Specify a time range for which to obtain user statistics.
The API expects a string starting with a capitalised time unit (\code{"DAYS"}, \code{"HOURS"}),
followed by a dash ("-") and an integer value specifying the length of the period.
Default is \code{"DAYS-31"}.}

\item{resolution}{Specifying a temporal resolution for the user statistics.
should be one of \code{"DAILY"} (default), \code{"MONTHLY"}, or \code{"HOURLY"}.}
}
\value{
A \code{data.frame} with requested information for the user associated with
the provided \code{token}.
}
\description{
In order to guarantee good performance for all users, the Sentinel Hub applies
\href{https://docs.sentinel-hub.com/api/latest/api/overview/rate-limiting/}{rate limiting}.
This policy enforces monthly quotas to your usage. To check your quota and current
usage, you can call \code{dse_usage()} or \code{dse_user_statistics()}.
}
\examples{
if (interactive() && dse_has_client_info()) {
  dse_usage()
  dse_user_statistics()
}
}
\seealso{
\code{\link[=dse_access_token]{dse_access_token()}}
}
