% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odata_products.R
\name{dse_odata_download_path}
\alias{dse_odata_download_path}
\title{TODO}
\usage{
dse_odata_download_path(
  product,
  destination,
  compressed = TRUE,
  ...,
  token = dse_access_token()
)
}
\arguments{
\item{product}{TODO}

\item{destination}{TODO}

\item{compressed}{A \code{logical} value. If set to\code{TRUE} (default), the product
will be downloaded as a zipped archive file.}

\item{...}{TODO}

\item{token}{TODO}
}
\description{
TODO Often doesn't work
}
\examples{
if (interactive() && dse_has_client_info()) {
  ##TODO examples not always working
  dse_odata_download_path(
    "85d8fe9d-cf8e-4c51-b4fd-7b811b514673",
    tempfile(fileext = ".nc"), compressed = FALSE)

  dse_odata_download_path(
    "002f0c9e-8a4c-465b-9e03-479475947630",
    tempfile(fileext = ".zip"))
}
}
