% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinelhub.R
\name{dse_sh_collections}
\alias{dse_sh_collections}
\title{List Sentinel Hub Collections}
\usage{
dse_sh_collections(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
Returns a \code{data.frame} with information about the
collections available from the Sentinel Hub
}
\description{
List collections that are available from the Sentinel Hub.
}
\examples{
if (interactive()) {
  dse_sh_collections()
}
}
