% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinelhub.R
\name{dse_sh_prepare_input}
\alias{dse_sh_prepare_input}
\alias{dse_sh_prepare_output}
\title{Prepare Input and Output Fields for Sentinel Hub Request}
\usage{
dse_sh_prepare_input(
  bounds,
  time_range,
  collection_name = "sentinel-2-l2a",
  id = NA,
  max_cloud_coverage = 100,
  mosaicking_order = "default",
  upsampling = "default",
  downsampling = "default",
  harmonize_values = FALSE,
  ...
)

dse_sh_prepare_output(
  width = 512,
  height = 512,
  output_identifier = "default",
  output_format = "image/tiff",
  bbox,
  ...
)
}
\arguments{
\item{bounds}{TODO}

\item{time_range}{TODO}

\item{collection_name}{TODO}

\item{id}{TODO}

\item{max_cloud_coverage}{Maximum cloud cover to be included in the
process. Value between 0 and 100 (default) percent.}

\item{mosaicking_order}{TODO}

\item{upsampling}{TODO}

\item{downsampling}{TODO}

\item{harmonize_values}{TODO}

\item{...}{TODO}

\item{width}{TODO}

\item{height}{TODO}

\item{output_identifier}{TODO}

\item{output_format}{TODO}

\item{bbox}{description}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
library(sf)

shape <- st_bbox(c(xmin = 5.261, ymin = 52.680,
                   xmax = 5.319, ymax = 52.715), crs = 4326) |>
           st_as_sfc()
dse_sh_prepare_input(
  bounds = c(5.261, 52.680, 5.319, 52.715),
  time_range = c("2025-06-01 UTC", "2025-07-01 UTC")
)
dse_sh_prepare_input(
  bounds = shape,
  time_range = c("2025-06-01 UTC", "2025-07-01 UTC")
)

dse_sh_prepare_output(bbox = shape)
}
\references{
\itemize{
\item \url{https://apps.sentinel-hub.com/requests-builder/}
}
}
\seealso{
\code{\link[=dse_sh_process]{dse_sh_process()}}
}
