% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinelhub.R
\name{dse_sh_process}
\alias{dse_sh_process}
\title{Process Satellite Data and Download Result}
\usage{
dse_sh_process(
  input,
  output,
  evalscript,
  destination,
  ...,
  token = dse_access_token()
)
}
\arguments{
\item{input}{TODO}

\item{output}{TODO}

\item{evalscript}{TODO}

\item{destination}{A file name to store the downloaded image.}

\item{...}{TODO}

\item{token}{For authentication, many of the Dataspace Ecosystem uses
an access token. Either provide your access token, or obtain one automatically
with \code{\link[=dse_access_token]{dse_access_token()}} (default). Without a valid token you will likely get
an "access denied" error.}
}
\value{
A \code{httr2_response} class object containing the
location of the downloaded file at its \code{destination}.
}
\description{
TODO
}
\examples{
input <- list(
  bounds = list(bbox = c(5.261, 52.680, 5.319, 52.715)),
  data = list(
    list(
      dataFilter = list(
        timeRange = list(from = "2025-06-21T00:00:00Z", to = "2025-07-21T00:00:00Z")
      ),
      type = "sentinel-2-l2a"
    )
  )
)
output <- list(
  width = 512, height = 515.09, responses = list(
    list(
      identifier = "default",
      format = list(type = "image/tiff")
    )
  )
)

evalscript <- dse_sh_get_custom_script("/sentinel-2/l2a_optimized/")

if (interactive() && dse_has_client_info()) {
  fl <- tempfile(fileext = ".tiff")
  dse_sh_process(input, output, evalscript, fl) #TODO
  
  if (requireNamespace("stars")) {
    library(stars)
    enkhuizen <- read_stars(fl) |> suppresWarnings()
    plot(enkhuizen, rgb = 1:3, axes = TRUE, main = "Enkhuizen")
  }
}
}
