% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac.R
\name{dse_stac_client}
\alias{dse_stac_client}
\title{Obtain Information About the STAC Client}
\usage{
dse_stac_client(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
Returns a \code{data.frame} with the requested information
}
\description{
Returns information about the STAC client used in the Data Space Ecosystem
}
\examples{
if (interactive()) {
  dse_stac_client()
}
}
