% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac.R
\name{dse_stac_collections}
\alias{dse_stac_collections}
\title{TODO}
\usage{
dse_stac_collections(...)
}
\arguments{
\item{...}{Ignored}
}
\description{
TODO
}
\examples{
if (interactive()) {
  dse_stac_collections()
}
}
