% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac.R
\name{dse_stac_queryables}
\alias{dse_stac_queryables}
\title{Get Queryables for a STAC Collection}
\usage{
dse_stac_queryables(collection, ...)
}
\arguments{
\item{collection}{Name of the collection for which to get the queryables.}

\item{...}{Ignored}
}
\value{
Returns a named list with information about elements that can be used
to query the \code{collection}
}
\description{
When searching through a collection with \code{\link[=dse_stac_search_request]{dse_stac_search_request()}}, it
can be helpful to know which elements can be used to filter the search
results (using \code{\link[dplyr:filter]{dplyr::filter()}}). Calling \code{\link[=dse_stac_queryables]{dse_stac_queryables()}} tells
you which aspects are available for querying and expected formats.
}
\examples{
if (interactive()) {
  dse_stac_queryables("sentinel-1-grd")
}
}
