% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac.R
\name{dse_stac_search_request}
\alias{dse_stac_search_request}
\title{TODO}
\usage{
dse_stac_search_request(collections, ids, ...)
}
\arguments{
\item{...}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO
library(dplyr)

if (interactive()) {
  dse_stac_search_request() |>
    collect()
}
}
